refactor: data/db/entity 패키지 model로 이름 변경
