+++ app/src/main/java/woowacourse/shopping/data/db/model/ProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.db.entity
+package woowacourse.shopping.data.db.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
